PC-1008 Sqoop processors correctly report record counts (import, export) exceeding Integer limits.

+++ integrations/nifi/nifi-nar-bundles/nifi-hadoop-bundle/nifi-hadoop-processors/src/main/java/com/thinkbiganalytics/nifi/v2/sqoop/core/ExportSqoop.java
@@ -312,7 +312,7 @@ public class ExportSqoop extends AbstractNiFiProcessor {
 
         int resultExportStatus = sqoopExportProcessResult.getExitValue();
         SqoopUtils sqoopUtils = new SqoopUtils();
-        int recordsExportCount = sqoopUtils.getSqoopExportRecordCount(sqoopExportProcessResult, logger);
+        long recordsExportCount = sqoopUtils.getSqoopExportRecordCount(sqoopExportProcessResult, logger);
 
         String sqoopExportCommandWithCredentialsMasked = sqoopUtils.maskCredentials(sqoopExportCommand,
                                                                                     sqoopUtils.getCredentialsToMask());
